:herb: splits functions/plugin tests
Include tests for c1686ac.
